Separate table partitioning logic into separate module
